Desafio Previred: Buenas practicas; se agreaga Junit y pruebas unitarias. Uso correcto de datos y ajustes en el manejo de excepciones

+++ src/main/java/cl/previred/desafio/dto/EmpleadoRequest.java
@@ -1,13 +1,15 @@
 package cl.previred.desafio.dto;
 
+import java.math.BigDecimal;
+
 public class EmpleadoRequest {
     private String nombre;
     private String apellido;
     private String rut;
     private String cargo;
-    private Double salario;
-    private Double bono;
-    private Double descuentos;
+    private BigDecimal salario;
+    private BigDecimal bono;
+    private BigDecimal descuentos;
 
     public EmpleadoRequest() {
     }
@@ -44,27 +46,27 @@ public class EmpleadoRequest {
         this.cargo = cargo;
     }
 
-    public Double getSalario() {
+    public BigDecimal getSalario() {
         return salario;
     }
 
-    public void setSalario(Double salario) {
+    public void setSalario(BigDecimal salario) {
         this.salario = salario;
     }
 
-    public Double getBono() {
+    public BigDecimal getBono() {
         return bono;
     }
 
-    public void setBono(Double bono) {
+    public void setBono(BigDecimal bono) {
         this.bono = bono;
     }
 
-    public Double getDescuentos() {
+    public BigDecimal getDescuentos() {
         return descuentos;
     }
 
-    public void setDescuentos(Double descuentos) {
+    public void setDescuentos(BigDecimal descuentos) {
         this.descuentos = descuentos;
     }
 }
